requirements.txt
PyYAML>=6.0
pytest>=7.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
